+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## UNRELEASED
+
+* new helper `m/decode-response-body` to return response body, decoded based on response `Content-Type` header. Throws if the body can't be parsed.
+
+```clj
+(-> {:headers {"accept" "application/edn"}}
+    ((middleware/wrap-format
+       (constantly {:status 200, :body {:date (Date. 0)}})))
+    (m/decode-response-body))
+; {:date #inst"1970-01-01T00:00:00.000-00:00"}
+```
+
 ## 0.6.2 (9.12.2018)
 
 * Added 2-arity to `encode` & `decode`, using the default instance, fixes [#86](https://github.com/metosin/muuntaja/issues/86), thanks to [valerauko](https://github.com/valerauko).